fix: localize dst config options in details
